simple-agent/pyproject.toml
[project]
name = "simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-mcp-adapters>=0.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "ruff>=0.15.4",
]
